Utils update (#6)
flake.nix changes:
Added ```packageDefinitions``` to the list of things output by this flake with ```inherit packageDefinitions;```
This should allow for much less copy pasting when using customPackager
This is a 1 line addition. It is the only important change to flake.nix

Other miscellaneous comment additions to flake.nix, as well as 1 change that illustrates the nuances of the nix ```with pkgs.something``` syntax but changes nothing.

builder/utils file:

Added 1 util to the exported utils set, allowing for easy creation of an overlay containing multiple neovim packages.

Made ```utils.mkOverlays``` into a composition of 2 other functions, and also added them to the utils set already being exported.

Comment additions

Other than that, the only things changed were improvements and additions to readme and help.