Added an example of using pckr to download plugins instead of nix from the same file. (#12)
pckr example is included, 0 nix changes were required, only an extra lua function or 2.

Help added.

All plugins and lsps do work, HOWEVER,

since you arent downloading via nix, you are in charge of dependencies now. Have fun!

But it does actually work great if you get that right. dependencies for nixCats-nvim as-is are included in help file, and in the lua file for pckr itself.