Feature: allowed choice of suffix or prefix for path and LD
Now can set the default behavior of sharedLibraries and lspsAndDeps
sections via the settings set for a package. More atomic control
of this behavior is to be done via extraWrapperArgs section, as these
sections are effectively just wrappers for extraWrapperArgs,
but being able to change the default seemed desirable without being
complicated for the user.

ran nix flake update

formatting