now wont throw an error when you have no init.lua at root
(i.e. it wont throw if no config) and it can also
accept an init.vim in its place.

To be completely honest, I somehow never thought about this being
different. In my brain originally, my thoughts were:
"oh you have no config of course theres an error"
but no, it should not be like that.

Also, this is not in any way a breaking change. So I may as well do it.