updated the kickstart-nvim template example
kickstart-nvim template, designed to be a tutorial on using the lazy
wrapper, has been updated to the newest version of kickstart-nvim!

It also implements the whole thing now, not just the main init.lua

I also made a slight update to the lazy.nvim wrapper for cleanliness,
but I did not change any of its behavior.