NixCats plugin user command improvements and naming consistency.
Note: some deprecations have been made. Warnings have been provided that the removal of the old version of these commands is set for 2024-09-01

Combined all the debug info user commands into the single NixCats command, which has been given arguments (and autocomplete for those arguments)

Made the naming scheme of the user commands and things in nixCats plugin match the name of the value in the nixCats plugin.

This means all remaining instances of included in the nixCats plugins have been replaced by pawsible

No further breaking changes are planned. The design of nixCats allows them to be rare, as most things can simply be added as new options or categories that can be simply ignored.

However, this was necessary for fixing the original ad-hoc naming scheme of things within the nixCats plugin, which was confusing. They are now much more consistent.

In exchange, I give you a much more capable command for obtaining debug information, and time to adjust.

Apologies for the changes, but hopefully you will enjoy the improved user command!