Added NVIM_WRAPPER_PATH_NIX env var Added an environment variable that contains the wrapper script's absolute path.
This is useful for things like vim-startuptime that need to know this
information, and would have otherwise used vim.v.progpath which
contains the path to the actual nvim binary being wrapped, which would
then not include your configuration.

Previously the only way to get this info was
vim.fn.exepath(nixCats('nixCats_packageName'))
And it only worked if the package was installed to the path.

If you wish to change vim.v.progpath, you can set it to equal this
variable yourself. I am unsure if this would cause issues with some
plugins, so you must do this yourself if you want that.

I also added a field in luaUtils template for getting this value,
or defaulting to vim.v.progpath if not set, such as if the configuration
was loaded without using nix.