refactor for future-proofing
main builder lost the (undocumented) ability to take a resolved
pkgs object instead of nixpkgs + system +
extra_pkg_config + dependencyOverlays

The ability to do so was simply an unnecessary side effect of the way
the modules were generated, which has been fixed.

In exchange, updates in nixpkgs are even less likely to break things,
and the builder is once again allowed to access the channel source,
giving increased freedom for future work if required.