mkNixosModules and mkHomeModules: all args now optional
No behavior changes, if defaultPackageName is passed, the
moduleNamespace will still be `[defaultPackageName]`

But if you don't pass either defaultPackageName or moduleNamespace,
the moduleNamespace will be `[ "nixCats" ]`

This ensures there are no behavior changes to previously valid useage,
only the ability to pass no arguments at all
to mkNixosModules or mkHomeModules has been added.

Added the moduleNamespace argument to mkNixosModules and mkHomeModules
in the templates to aid in clarity for new users.

Also improved formatting of the generated utils docs