feat(overrides and pkgs): see below (#132)
Added `.overrideNixCats` alias for `.override`
to avoid it being shadowed by `callPackage`

Allowed overriding of select passthru values via overrideAttrs
to change the args to the builder function.

allowed pkgs to be passed to main builder function
instead of nixpkgs and system arguments
and inherit overlays and other attributes from it.

This allows flakeless users to avoid reinstantiating nixpkgs.

deprecated being able to pass a system wrapped set of lists instead of
just a list as dependencyOverlays, something that is not recommended,
as overlays already can access the system variable internally.

Co-authored-by: BirdeeHub <birdee@localhost>