feat(wrappedCfgPath): new setting and args for packageDefinitions
https://nixcats.org/nixCats_format.html#nixCats.flake.outputs.packageDefinitions

https://nixcats.org/nixCats_format.html#nixCats.flake.outputs.settings

added 2 arguments to packageDefintions, `luaPath` and `this`

added wrappedCfgPath setting, for when wrapRc == true:

if not null, will replace the main luaPath as the wrapped config path
This may be used to apply optional build steps to your config directory
per package.
you may get the original `luaPath` in the arguments to your package
definition.
if it is a function, it will recieve the same
arguments as packageDefinitions and should return the new path